net/mlx4_core: Keep VLAN/MAC tables mirrored in multifunc HA mode
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Moni Shoua <monis@mellanox.com>
commit 5f61385d2ebc2bd62bc389c7da0d8d2f263be1eb

Due to HW limitations, indexes to MAC and VLAN tables are always taken
from the table of the actual port. So, if a resource holds an index to
a table, it may refer to different values during the lifetime of the
resource,  unless the tables are mirrored. Also, even when
driver is not in HA mode the policy of allocating an index to these
tables is such to make sure, as much as possible, that when the time
comes the mirroring will be successful. This means that in multifunction
mode the allocation of a free index in a port's table tries to make sure
that the same index in the other's port table is also free.

	Signed-off-by: Moni Shoua <monis@mellanox.com>
	Reviewed-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
	Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 5f61385d2ebc2bd62bc389c7da0d8d2f263be1eb)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>