IB/mlx4: Use the VF base-port when demuxing mad from wire
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Or Gerlitz <ogerlitz@mellanox.com>
commit f1b4e12a9ab6cc08b1a047bf021d90c7e07b1517

Under HA mode, it's possible that the VF registered its GID
(and expects to get mads through the PV scheme) on a port which is
different from the one this mad arrived on, due to HA fail over.

Therefore, if the gid is not matched on the port that the packet arrived
on, check for a match on the other port if HA mode is active -- and if a
match is found on the other port, continue processing the mad using that
other port.

	Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
	Reviewed-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit f1b4e12a9ab6cc08b1a047bf021d90c7e07b1517)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>