net/mlx5_core: Connect flow tables
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Maor Gottlieb <maorg@mellanox.com>
commit f90edfd279f35a5c62003eeee107f03f3b1544bc

Flow tables from different priorities should be chained together.
When a packet arrives we search for a match in the
by-pass flow tables (first we search for a match in priority 0
and if we don't find a match we move to the next priority).
If we can't find a match in any of the bypass flow-tables, we continue
searching in the flow-tables of the next priority, which are the
kernel's flow tables.

Setting the miss flow table in a new flow table to be the next one in
the list is performed via create flow table API. If we want to change an
existing flow table, for example in order to point from an
existing flow table to the new next-in-list flow table, we use the
modify flow table API.

	Signed-off-by: Maor Gottlieb <maorg@mellanox.com>
	Signed-off-by: Moni Shoua <monis@mellanox.com>
	Signed-off-by: Matan Barak <matanb@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit f90edfd279f35a5c62003eeee107f03f3b1544bc)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>