IB/core: Support accessing SA in virtualized environment
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Eli Cohen <eli@mellanox.com>
commit a0c1b2a3508714281f604db818fa0cc83c2f9b6a

Per the ongoing standardisation process, when virtual HCAs are present
in a network, traffic is routed based on a destination GID. In order to
access the SA we use the well known SA GID.

We also add a GRH required boolean field to the port attributes which is
used to report to the verbs consumer whether this port is connected to a
virtual network. We use this field to realize whether we need to create
an address vector with GRH to access the subnet administrator. We clear
the port attributes struct before calling the hardware driver to make
sure the default remains that GRH is not required.

	Signed-off-by: Eli Cohen <eli@mellanox.com>
	Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>
	Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit a0c1b2a3508714281f604db818fa0cc83c2f9b6a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>