enic: Update driver to use __dev_uc/mc_sync/unsync calls
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Alexander Duyck <alexander.h.duyck@intel.com>
commit f009618ac57cd3feeb4622147e8d2e88e47b9c1c

This change updates the enic driver to make use of __dev_uc_sync and
__dev_mc_sync calls.  Previously the driver was doing its own list
management by storing the mc_addr and uc_addr list in a 32 address array.
With this change the sync data is stored in the netdev_addr_list structures
and instead we just track how many addresses we have written to the device.
When we encounter 32 we stop and print a message as occurred previously with
the old approach.

Other than the core change the only other bit needed was to propagate the
constant attribute with the MAC address as there were several spots where
is twas only passed as a u8 * instead of a const u8 *.

This patch is meant to maintain the original functionality without the use
of the mc_addr and uc_addr arrays.

	Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
	Acked-by: Govindarajulu Varadarajan <_govind@gmx.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit f009618ac57cd3feeb4622147e8d2e88e47b9c1c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>