netfilter: nf_tables: fix error handling of rule replacement
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Pablo Neira Ayuso <pablo@netfilter.org>
commit 59900e0a019e7c2bdb7809a03ed5742d311b15b3

In general, if a transaction object is added to the list successfully,
we can rely on the abort path to undo what we've done. This allows us to
simplify the error handling of the rule replacement path in
nf_tables_newrule().

This implicitly fixes an unnecessary removal of the old rule, which
needs to be left in place if we fail to replace.

	Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
(cherry picked from commit 59900e0a019e7c2bdb7809a03ed5742d311b15b3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>