KVM: x86: #PF error-code on R/W operations is wrong
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Nadav Amit <namit@cs.technion.ac.il>
commit c205fb7d7d4f81e46fc577b707ceb9e356af1456

When emulating an instruction that reads the destination memory operand (i.e.,
instructions without the Mov flag in the emulator), the operand is first read.
If a page-fault is detected in this phase, the error-code which would be
delivered to the VM does not indicate that the access that caused the exception
is a write one. This does not conform with real hardware, and may cause the VM
to enter the page-fault handler twice for no reason (once for read, once for
write).

	Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
	Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit c205fb7d7d4f81e46fc577b707ceb9e356af1456)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>