x86/mm/srat: Print non-volatile flag in SRAT
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
Rebuild_CHGLOG: - [x86] mm/srat: Print non-volatile flag in SRAT (Jeff Moyer) [1345801 1348502 1348836 1350149 1350156]
Rebuild_FUZZ: 95.24%
commit-author Linda Knippers <linda.knippers@hp.com>
commit 31e09b18c863718939e3e9c30eee55f9011d85ee

With the addition of NVDIMM support, a question came up as to
whether NVDIMM ranges should be in the SRAT with this bit set.
I think the consensus was no because the ranges are in the NFIT
with proximity domain information there.

ACPI is not clear on the meaning of this bit in the SRAT.
If someone is setting it, we might want to ask them what they
expect to happen with it.

Right now this bit is only printed if all the ACPI debug
information is turned on.

	Signed-off-by: Linda Knippers <linda.knippers@hp.com>
	Acked-by: Thomas Gleixner <tglx@linutronix.de>
	Cc: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/20150901194154.GA4939@ljkz400
	Signed-off-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit 31e09b18c863718939e3e9c30eee55f9011d85ee)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>