memremap: fix highmem support
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-514.el7
commit-author Dan Williams <dan.j.williams@intel.com>
commit 182475b7a2831abf7e6ca83b2aced0bef5dcdfd3

Currently memremap checks if the range is "System RAM" and returns the
kernel linear address.  This is broken for highmem platforms where a
range may be "System RAM", but is not part of the kernel linear mapping.
Fallback to ioremap_cache() in these cases, to let the arch code attempt
to handle it.

Note that ARM ioremap will WARN when attempting to remap ram, and in
that case the caller needs to be fixed.  For this reason, existing
ioremap_cache() usages for ARM are already trained to avoid attempts to
remap ram.

The impact of this bug is low for now since the pmem driver is the only
user of memremap(), but this is important to fix before more conversions
to memremap arrive in 4.4.

	Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
	Reported-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
	Acked-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
	Signed-off-by: Dan Williams <dan.j.williams@intel.com>
(cherry picked from commit 182475b7a2831abf7e6ca83b2aced0bef5dcdfd3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>