blk-mq: improve warning for running a queue on the wrong CPU
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Jens Axboe <axboe@fb.com>
commit 0e87e58bf60edb6bb28e493c7a143f41b091a5e5

__blk_mq_run_hw_queue() currently warns if we are running the queue on a
CPU that isn't set in its mask. However, this can happen if a CPU is
being offlined, and the workqueue handling will place the work on CPU0
instead. Improve the warning so that it only triggers if the batch cpu
in the hardware queue is currently online.  If it triggers for that
case, then it's indicative of a flow problem in blk-mq, so we want to
retain it for that case.

	Signed-off-by: Jens Axboe <axboe@fb.com>
(cherry picked from commit 0e87e58bf60edb6bb28e493c7a143f41b091a5e5)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>